+++ package.json
@@ -21,6 +21,7 @@
     "@angular/common": "^15.0.0",
     "@angular/compiler": "^15.0.0",
     "@angular/core": "^15.0.0",
+    "@angular/elements": "^15.2.3",
     "@angular/forms": "^15.0.0",
     "@angular/material": "^15.2.2",
     "@angular/platform-browser": "^15.0.0",
