--- src/app/app-routing.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-const routes: Routes = [{
-  path: '',
-  loadChildren: () => import('./core/core.module').then(m => m.CoreModule)
-}];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
-})
-export class AppRoutingModule { }
