+++ src/app/app.component.html
@@ -1,2 +1,3 @@
-<router-outlet></router-outlet>
+<app-form></app-form>
+<app-drag-tool (start)="startService()" (stop)="stopService()" [isVRRcord]="isVRStart"></app-drag-tool>
