+++ src/app/app.component.ts
@@ -1,5 +1,6 @@
-import { ChangeDetectionStrategy, Component, NgZone } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, HostListener, NgZone } from '@angular/core';
 import { Observable } from 'rxjs';
+import { VoiceRecognitionServiceService } from './services/voice-recognition-service.service';
 
 @Component({
   selector: 'app-root',
@@ -10,57 +11,171 @@ import { Observable } from 'rxjs';
 
 export class AppComponent {
 
+  @HostListener('window:click', ['$event'])
+  clickout(event: any) {
+    console.log(event, 'event12');
+    this.selectIdByEvent(event);
+  }
+
+  @HostListener('keyup', ['$event'])
+  onKeyDown(e: any) {
+    if ((e.shiftKey && e.keyCode == 9) || e.keyCode == 9) {
+      console.log('shift and tab');
+      this.selectIdByEvent(e);
+    } else if (e.keyCode == 27) {
+      this.isVRStart = !this.isVRStart;
+      this.isVRStart ? this.startService() : this.stopService();
+    }
+  }
+
   title = 'IDP-Speech-To-Text';
   boo = false;
   speech: string = '';
   voice = ''
+  selectedId = '';
+  isVRStart: boolean = false;
 
-  constructor(private _ngZone: NgZone) {
+  constructor(private _ngZone: NgZone, public voiceRecognitionServiceService: VoiceRecognitionServiceService, private cdRef: ChangeDetectorRef) {
 
   }
 
-  // getTranscript({ locale = 'en-US' }: { locale?: string } = {}): Observable<string> {
-  //   return new Observable(observer => {
-  //     const SpeechRecognition = window.SpeechRecognition || webkitSpeechRecognition;
-  //     const SpeechGrammarList = window.SpeechGrammarList || webkitSpeechGrammarList;
-  //     const SpeechRecognitionEvent = window.SpeechRecognitionEvent || webkitSpeechRecognitionEvent;
+  selectIdByEvent(event: any) {
+    let element = event.target || event.srcElement || event.currentTarget;
+    // Get the id of the source element
+    let elementId = element.id;
+    console.log(elementId, 'elementId');
+    if (elementId.startsWith('vr-ans')) {
+      this.selectedId = elementId;
+    }
+  }
 
-  //     // const SpeechRecognition = window['webkitSpeechRecognition'];
-  //     const speechRecognition = new SpeechRecognition();
-  //     speechRecognition.continuous = true;
-  //     speechRecognition.interimResults = true;
-  //     speechRecognition.lang = locale;
-  //     speechRecognition.onresult = (speechRecognitionEvent) => {
-  //       var interim_transcript = '';
-  //       for (var i = speechRecognitionEvent.resultIndex; i < speechRecognitionEvent.results.length; ++i) {
-  //         if (speechRecognitionEvent.results[i].isFinal) {
-  //           this.boo = true;
-  //           this._ngZone.run(() => observer.next(speechRecognitionEvent.results[i][0].transcript.trim()));
-  //         }
-  //         else {
-  //           this.boo = false;
-  //           interim_transcript += speechRecognitionEvent.results[i][0].transcript;
-  //           this._ngZone.run(() => observer.next(interim_transcript.trim()));
-  //         }
+  startService() {
+    this.isVRStart = true;
+    this.recognize();
+  }
 
-  //       }
-  //     };
-  //     speechRecognition.start();
+  stopService() {
+    this.isVRStart = false;
+    this.voiceRecognitionServiceService.stop();
+  }
 
-  //     return () => speechRecognition.abort();
-  //   });
-  // }
+  recognize() {
+    this.voiceRecognitionServiceService.getTranscript()
+      .subscribe(transcript => {
+        console.log(transcript, "res")
+        if (transcript !== '' && this.voiceRecognitionServiceService.boo) {
+          this.voice = this.voice + ' ' + transcript;
+        }
+        else {
+          this.speech = transcript;
+          console.log(this.selectedId, 'selectedId');
+          let inputElement: any = document.getElementById(this.selectedId) as HTMLInputElement | null;
+          console.log(inputElement?.parentElement.id, 'idtest');
+          // console.log(test?.options,'options123');
+          // console.log(test?.options?.value,'options123456');
+          //console.log(test?.options?.textContent,'options123456');
+          if (inputElement.type == "select-one") {
+            this.selectDropDownValue(inputElement?.options, transcript?.toLocaleLowerCase())
+          } else if (inputElement.type == "radio") {
+            console.log(inputElement.name, 'name12');
+            this.selectRadioValue(inputElement.name, transcript?.toLocaleLowerCase());
+          } else if (inputElement.type == 'checkbox') {
+            let splitval = 'vr-div-' + this.selectedId.split('-')[2];
+            this.selectCheckBoxValue(splitval, transcript?.toLocaleLowerCase());
+          } else if (inputElement.type == 'datetime-local') {
+            this.selectDateTime(transcript);
+          }
+          else {
+            inputElement.value = transcript?.toLocaleLowerCase();
+          }
+        }
+        this.cdRef.detectChanges();
+      });
+  }
 
-  // recognize() {
-  //   this.getTranscript()
-  //     .subscribe(transcript => {
-  //       if (transcript !== '' && this.boo) {
-  //         this.voice = this.voice + ' '+ transcript;
-  //       }
-  //       else
-  //       {
-  //         this.speech = transcript
-  //       }
-  //     });
-  // }
+  selectDateTime(value: any) {
+    let splitvalue = value.split(' ');
+    console.log(splitvalue.length, 'length');
+    if (splitvalue.length == 3) {
+      let test: any = document.getElementById(this.selectedId) as HTMLInputElement | null;
+      test.value = this.datetimeLocal(splitvalue[0] + '-' + splitvalue[1] + '-' + splitvalue[2]);
+    }
+  }
+
+  datetimeLocal(datetime: any) {
+    console.log(datetime, 'datetime')
+    const dt = new Date(datetime);
+    dt.setMinutes(dt.getMinutes() - dt.getTimezoneOffset());
+    console.log(dt.toISOString().slice(0, 16), 'datetime123')
+    return dt.toISOString().slice(0, 16);
+  }
+
+  selectCheckBoxValue(parentId: any, value: any) {
+    // console.log(parentId,'parentId');
+    var stringSimilarity = require("string-similarity");
+    var itemForm: any = document.getElementById(parentId) as HTMLInputElement | null;; // getting the parent container of all the checkbox inputs
+    var checkBoxes = itemForm.querySelectorAll('input[type="checkbox"]');
+    // console.log(checkBoxes,'checkBoxes');
+    for (var i = 0; i < checkBoxes.length; i++) {
+      var selector = 'label[for=' + checkBoxes[i].id + ']';
+      var label: any = document.querySelector(selector) as HTMLInputElement | null;
+      // console.log(label.innerHTML,'innerHTML');
+      let matchPer = stringSimilarity.compareTwoStrings(value, label.innerHTML?.toLocaleLowerCase());
+      console.log(label.innerHTML?.toLocaleLowerCase(), '-', value, '-', matchPer, 'txcont');
+      if (matchPer >= 0.8) {
+        let test: any = document.getElementById(checkBoxes[i].id) as HTMLInputElement | null;
+        console.log(test.checked, 'test.checked');
+        test.checked = true;
+      }
+      //var text = label.innerHTML;
+      // do stuff
+    }
+  }
+
+  selectRadioValue(name: any, value: any) {
+    let radios = document.getElementsByName(name);
+    var stringSimilarity = require("string-similarity");
+    console.log(radios, 'radios');
+    for (var i = 0; i < radios.length; i++) {
+      var selector = 'label[for=' + radios[i].id + ']';
+      var label: any = document.querySelector(selector) as HTMLInputElement | null;
+      // console.log(label.innerHTML,'innerHTML');
+      let matchPer = stringSimilarity.compareTwoStrings(value, label.innerHTML?.toLocaleLowerCase());
+      console.log(label.innerHTML?.toLocaleLowerCase(), '-', value, '-', matchPer, 'txcont');
+      if (matchPer >= 0.8) {
+        let test: any = document.getElementById(radios[i].id) as HTMLInputElement | null;
+        test.checked = true;
+      }
+      //var text = label.innerHTML;
+      // do stuff
+    }
+  }
+
+  selectDropDownValue(options: any, value: any) {
+    var stringSimilarity = require("string-similarity");
+
+    console.log(value, 'value123');
+    for (var i = 0; i < options.length; i++) {
+      let ele = options[i];
+      // options.forEach((ele:any) => {
+      let matchPer = stringSimilarity.compareTwoStrings(value, ele.textContent?.toLocaleLowerCase());
+      console.log(ele.textContent, '-', value, 'txcont');
+      console.log(matchPer, 'txvalue');
+      if (matchPer >= 0.8) {
+        console.log(ele.value, 'txvalue123');
+        let test: any = document.getElementById(this.selectedId) as HTMLInputElement | null;
+        test.value = ele.value;
+      }
+    }
+    // });
+
+  }
+
+  cancel() {
+
+  }
+
+  signUp() {
+
+  }
 }
