+++ src/app/components/drag-tool/drag-tool.component.spec.ts
