+++ src/app/components/form/form.component.html
@@ -125,4 +125,4 @@
 </div>
 
 
-<app-drag-tool (start)="startService()" (stop)="stopService()" [isVRRcord]="isVRStart"></app-drag-tool>
+<!-- <app-drag-tool (start)="startService()" (stop)="stopService()" [isVRRcord]="isVRStart"></app-drag-tool> -->
