+++ src/app/components/recognition-tester/recognition-tester.component.spec.ts
