--- src/app/core/components/speech-to-text/speech-to-text.component.html
@@ -1,12 +1,0 @@
-<!-- <div class="wrapper">
-  <div class="header-cnt">
-    <h1>Voice recognition</h1>
-  </div>
-  <div class="controller-cnt">
-    <button style="margin: 10px;" (click)="startService()">Start</button>
-    <button style="margin: 10px;" (click)="stopService()">Stop</button>
-  </div>
-  <div class="result-cnt">
-    <p>{{voiceRecognitionServiceService.text}}</p>
-  </div>
-</div> -->
