--- src/app/core/components/speech-to-text/speech-to-text.component.scss
@@ -1,31 +1,0 @@
-.wrapper {
-  margin: 20px;
-  border-radius: 2px;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  .header-cnt {
-    align-items: center;
-    justify-content: center;
-  }
-
-  .controller-cnt {
-    button{
-      padding: 10px;
-      font-weight: 600;
-      &:hover{
-        size: 100%;
-      }
-    }
-  }
-
-  .result-cnt {
-    min-width: 100%;
-    min-height: 300px;
-    border: 1px solid gray;
-    background: gray;
-    border-radius: 4PX;
-  }
-}
