--- src/app/core/core-routing.module.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { FormComponent } from './components/form/form.component';
-import { SpeechToTextComponent } from './components/speech-to-text/speech-to-text.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: FormComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-
-export class CoreRoutingModule { }
