--- src/app/core/core.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { CoreRoutingModule } from './core-routing.module';
-import { SharedModule } from '../shared/shared.module';
-import { FormComponent } from './components/form/form.component';
-
-@NgModule({
-  declarations: [
-    FormComponent
-  ],
-  imports: [
-    CommonModule,
-    CoreRoutingModule,
-    SharedModule
-  ],
-  exports:[
-  ]
-})
-
-export class CoreModule { }
