+++ src/index.html
@@ -11,6 +11,7 @@
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 </head>
 <body class="mat-typography">
-  <app-root></app-root>
+  <!-- <app-root></app-root> -->
+  <voice-reg-tool></voice-reg-tool>
 </body>
 </html>
